requirements.txt
anyio==3.7.0
astroid==3.0.2
bcrypt==4.1.2
black==23.3.0
cfgv==3.3.1
click==8.1.3
dill==0.3.7
distlib==0.3.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.2
fastapi==0.99.1
filelock==3.12.2
greenlet==3.0.3
h11==0.14.0
identify==2.5.24
idna==3.4
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.8.0
pre-commit==3.3.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pydantic==1.10.10
pydentic==0.0.1.dev3
pylint==3.0.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-stdnum==1.18
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.7.1
uvicorn==0.22.0
virtualenv==20.23.1
